scicos: fix some compilation warnings

+++ scilab/modules/scicos/sci_gateway/cpp/sci_scicos_debug.cpp
@@ -82,7 +82,7 @@ types::Function::ReturnValue sci_scicos_debug(types::typed_list &in, int _iRetCo
             return types::Function::Error;
         }
 
-        C2F(cosdebug).cosd = pIn->get(0);
+        C2F(cosdebug).cosd = (int) pIn->get(0);
     }
 
     return types::Function::OK;
